requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backoff==2.2.1
beanie==1.20.0
beautifulsoup4==4.12.2
bs4==0.0.1
cairocffi==1.6.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
discord.py==2.3.1
dnspython==2.4.0
frozenlist==1.4.0
h11==0.14.0
html2image==2.0.3
httpcore==0.17.3
idna==3.4
lazy-model==0.0.5
markdownify==0.11.6
motor==3.2.0
multidict==6.0.4
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.11
pymongo==4.4.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
toml==0.10.2
topggpy==1.4.0
typing_extensions==4.7.1
urllib3==1.26.18
w3lib==2.1.2
webencodings==0.5.1
websockets==10.4
yarl==1.9.2
